♻️ Extract run.go
